refactor: move `BetterAuthAdvancedOptions` to core
This moves the `BetterAuthAdvancedOptions` type and related interfaces
to the core package, making them available for plugins and other packages
to extend functionality.
