+++ packages/core/src/index.ts
@@ -1,1 +1,3 @@
-export {};
+// Shared types and interfaces for BetterAuth, allowing plugins to extend functionality.
+export interface BetterAuthMutators<O, C> {}
+export * from "./types";
