+++ packages/core/src/types/helper.ts
@@ -1,1 +1,6 @@
+type Primitive = string | number | symbol | bigint | boolean | null | undefined;
+
 export type LiteralString = "" | (string & Record<never, never>);
+export type LiteralUnion<LiteralType, BaseType extends Primitive> =
+	| LiteralType
+	| (BaseType & Record<never, never>);
