+++ packages/core/src/types/index.ts
@@ -1,1 +1,2 @@
-export * from "./helper";
+export type * from "./helper";
+export type { BetterAuthAdvancedOptions, GenerateIdFn } from "./init-options";
